split pymatviz/structure_viz.py module into plotly, mpl and helpers submodules
