+++ pymatviz/structure_viz/__init__.py
@@ -1,0 +1,4 @@
+"""2D and 3D plots of Structures."""
+
+from pymatviz.structure_viz.mpl import structure_2d
+from pymatviz.structure_viz.plotly import structure_2d_plotly, structure_3d_plotly
